separate out plot functions and add to plot_interp
